Fix bug by capturing exception when shadow server is down
When server is down, HAProxy intercepts the request to show a nice "OUR
APPOLOGIES..." message. However, this causes an abrupt termination of
the https connection resulting in an exception in the `ping` function.
Instead of bubbling up, now the exception is caught and returns False
(since the exception only occurs when the server is down).