Simplify the BOLD_LIST_ORDER format
Manually specifying the order number in two places is error prone and
can easily be computed based on list index. It is better to let the
computer do it. The reason I didn't just do that from the beginning was
because I wasn't sure how the list was being used and specifying numbers
was a needed feature. It is not.