Dynamically add the `lib` import path
You'll also notice that I reordered the import statements of `XNAT_CHECK` and
`XNAT_GET` so that shared_values is imported first. This is because
shared_values has the side-effect of modifying the import path so that
the custom modules can be found on the import path. Otherwise you'll get
a ModuleNotFoundError.

Also, the variable `HCP_LIB_DIR` is renamed to `PYTHON_IMPORT_DIR`
because the latter name is more descriptive of how the value is actually used.