Add `slurm` version of script header (`slurm_head`)
This also removes the CHPC2 specific hardware from the slurm header
since it is not available in CHPC3 (yet).

Per Malcolm:
The third-generation cluster is based on a newer generation of CPUs, so there
are no Haswell CPUs.  Likewise, the GPUs are V100S so no K20X (to request a
GPU, specify the partition #SBATCH --partition=gpu
and request the resource #SBATCH --gres=gpu:1).

Eventually we may have to worry about these things when we incorporate the
2nd generation cluster into the 3rd, but that's a way off.