Remove ability to copy FS hierarchy, now can only symlink
Before PipelinePrereqDownloader had the ability to traverse a directory
and either copy files over (using `rsync -auL`) or make symlinks of the
files. This patch removes the ability to directly copy files over. This
feature was not being used in any pipeline and was therefore superfluous
code that would need to be maintained.

Additionally, the final result could be achieved by creating symlinks
and then running a script afterward that converts symlinks to real files.
This approach has the added benefit that usually only a few of the files
would need to be duplicates, the majority would still need to be
symlinks, so a precise conversion process would be speedier than a full
duplication.