Move MNINonLinear/T1w/T2w relocation code from `run_qunex.sh` to `XNAT_GET.py`
There is a step that the root directories MNINonLinear, T1w, and T2w, get moved into a different location required by qunex. Currently that step is being done in the `run_qunex.sh` script, which is a PROCESS-ing script. It makes more sense to keep all the filesystem preparation actions, like this one, in the GET step.