Verify resource deletion directly on filesystem
The previous assumption was that when the server request returned, the entire
resource was deleted. This has led to partial overwrite of the resource.
Having a mix of files from two different runs is bad. This commit checks
the filesystem directly to confirm that the deletion was complete before
moving forward. If 3 deletion requests are performed and the resource
still exists, a fatal error is thrown.