Use the protocol/port specified in `PUT_SERVER_LIST` (variables.yaml)
Before the protocol/port for the servers was hardcoded in the code. If in the
future, we want to revert to HTTPS, or mix some HTTP + HTTPS this would
be nontrivial. Instead, this lets us specify the full URI for each
server in one central location, the variables.yaml file.

This reverts commit b1b0e713f85cdc4b53de8a89ac2bb33e64a69728.