Add error check to PROCESS step
The QUNEX container does not exit with a non-zero exit status upon
error, instead it logs the error to STDOUT and creates error logs.
However, this behavior does not let SLURM know that the processing
failed. Therefore, downstream dependencies continue to run, even going
as far as putting the data to XNAT (via the PUT step). The only
indication users will have that the data didn't process correctly is the
lack of a `*.success` file.

Instead of waiting til the CHECK step to check for error logs, it is
better to check for the logs before exiting the PROCESS step. This will
cause SLURM job dependencies to stop.