Fix bug where starttime file is sometimes not generated
The error only occurs if `USE_SCRATCH_FOR_PROCESSING=True`. Because the
starttime file is trying to be created in the build space while the rest
of the files are in the scratch space.

This bug results in the CLEAN step not working since it relies on a
starttime file.