Use `logging` module to instead of `print()`
If using print, the messages get sent to stdout. However, they do not get
stored to the log file. This will result in missing messages when
debugging the log file. Also, it is nice to be able to specify `debug`,
`info`, `warning`, or `critical` levels for the messages. Messages at
the `debug` level will not show up in console (to not dilute important
messages).

Also fixing a minor issue with the directory path that gets printed to logs.
It has been describing the wrong path for the `STUDY_FOLDER` because
`STUDY_FOLDER` is actually inside `WORKING_DIR` not equal to.