Use `logging` module to instead of `print()`
If using print, the messages get sent to stdout. However, they do not get
stored to the log file. This will result in missing messages when
debugging the log file. Also, it is nice to be able to specify `debug`,
`info`, `warning`, or `critical` levels for the messages. Messages at
the `debug` level will not show up in console (to not dilute important
messages).